refactor: LottoNumbers을 더 General 하게 쓰도록 수정
