+++ src/test/java/lotto/LottoCalculatorTest.java
@@ -18,7 +18,7 @@ public class LottoCalculatorTest {
 	@Test
 	@DisplayName("1000원 미만일 때, 로또 구매 불가 테스트")
 	void purchaseMoneyErrorTest() {
-		assertThatThrownBy(() -> LottoCalculator.getLottoCount(LottoNumbers.LOTTO_PRICE - 1))
+		assertThatThrownBy(() -> LottoCalculator.getAvailableLottoNumbers(LottoNumbers.LOTTO_PRICE - 1))
 				.isInstanceOf(IllegalArgumentException.class)
 				.hasMessageContaining("1000원");
 	}
